Implemented authorization - Live coding session 5

+++ meetometer/Controllers/MeetingsController.cs
@@ -1,9 +1,11 @@
 ﻿using System.Web.Http;
 using meetometer.Repository;
 using meetometer.Repository.Entities;
+using Microsoft.AspNet.Identity;
 
 namespace meetometer.Controllers
 {
+    [Authorize]
     public class MeetingsController : ApiController
     {
         private IMeetingRepository repository;
@@ -15,16 +17,35 @@ namespace meetometer.Controllers
 
         public IHttpActionResult Get()
         {
-            return Ok(repository.GetAll());
+            return Ok(repository.GetAll(User.Identity.GetUserId()));
         }
 
         public IHttpActionResult Post([FromBody] Meeting meeting)
         {
+            var existing = repository.GetById(meeting.Id);
+            if (existing != null &&
+                existing.UserId != User.Identity.GetUserId())
+            {
+                return BadRequest();
+            }
+
+            meeting.UserId = User.Identity.GetUserId();
             return Ok(repository.SaveOrUpdate(meeting));
         }
 
         public IHttpActionResult Delete(int id)
         {
+            var existing = repository.GetById(id);
+            if (existing == null)
+            {
+                return NotFound();
+            }
+
+            if (existing.UserId != User.Identity.GetUserId())
+            {
+                return BadRequest();
+            }
+
             if (repository.Delete(id))
             {
                 return Ok();
